move functions to helper file
