+++ templates/magazine/magazine.js
@@ -2,10 +2,8 @@ import {
   getMetadata,
   createOptimizedPicture,
 } from '../../scripts/aem.js';
-import {
-  createElement,
-  getArticleTags,
-} from '../../scripts/common.js';
+import { getArticleTags } from '../../scripts/magazine-helper.js';
+import { createElement } from '../../scripts/common.js';
 
 async function buildArticleHero({ truckTags, categoryTag } = {}) {
   const title = getMetadata('og:title');
